+++ apps/baskets/models.py
@@ -1,17 +1,9 @@
 from django.contrib.auth import get_user_model
-from django.core.exceptions import ValidationError
 from django.db import models
-from django.shortcuts import get_object_or_404
 from django.utils.translation import gettext_lazy as _
 
 from apps.products.models import Product
-
-
-def validate_user_is_buyer(value):
-    """Валидация, является ли пользователь покупателем."""
-    user = get_object_or_404(get_user_model(), pk=value)
-    if not (user.personal or user.company.role == "customer"):
-        raise ValidationError(_("Only buyers can create baskets."))
+from apps.users.validators import validate_user_is_buyer
 
 
 class Basket(models.Model):
