+++ apps/orders/models.py
@@ -1,13 +1,12 @@
 from _decimal import ROUND_HALF_UP, Decimal
 from django.contrib.auth import get_user_model
 from django.core import validators
-from django.core.exceptions import ValidationError
 from django.db import models
-from django.shortcuts import get_object_or_404
 from django.utils.translation import gettext_lazy as _
 
 from apps.core.models import BaseModel
 from apps.products.models import Product
+from apps.users.validators import validate_user_is_buyer
 
 
 class OrderManager(models.Manager):
@@ -18,13 +17,6 @@ class OrderManager(models.Manager):
             .prefetch_related("orders__product__user__company", "orders__product__images")
             .order_by("-created_at")
         )
-
-
-def validate_user_is_buyer(value):
-    """Валидация, является ли пользователь покупателем."""
-    user = get_object_or_404(get_user_model(), pk=value)
-    if not (user.personal or user.company.role == "customer"):
-        raise ValidationError(_("Only buyers can create orders."))
 
 
 class Order(BaseModel):
